Working script creating one file at a time
Original design was to read in a Rx file (name, dose, quantity) and zipcode file, then construct the URLS based on the Rx file and cycle through the URLS and zipcodes.  As I was working on this, goodrx started to throw an error saying it detected webscraping.  To not put my project at risk for late completion, I opted to run the script separately for each URL and zipcode combination to make sure I was not banned from the site before getting the data I need.  But, left in the loop structure so that I can return to this later to try to fix wait times, etc., so as not to be detected when running as one large script.